requirements.txt
numpy
Pillow
tensorflow
scikit-learn
h5py
